What_is_mem/api/v1/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User

class UserLoginSerialize(serializers.ModelSerializer):

    class Meta():
        model = User
        field = ["id", "username", "password"]