requirements.txt
Django~=4.2
mysqlclient
mysql-connector-python
Pillow
aiohttp~=3.8.4
aiohttp-jinja2
aiohttp-session
djangorestframework